refactor: move _AnyToExistsTransform to `analysis.py`
